Adds the functionality to hold values constant in a circuit
When defining a circuit you can now pass a constants dictionary which
specifies the values to hold constant during the fit by name. The initial
guess list can either contain None as a placeholder for the constant or
be shortened by the number of constants that are passed. Constants are
also displayed separately from initial guesses and fit parameters when the
circuit is printed.

Related to ECSHackWeek/impedance.py/#29 and ECSHackWeek/impedance.py/#34